Initialise query to prevent NullPointerException
Instead of setting query to null, set it to an empty string. Otherwise your 'query.replaceAll(" ", "%20")' call in createURL() will result in a NullPointerException, when the user hasn't set a query (cause they aren't looking for a specific title). You could also just add a check if query is null, but that's more code than my solution 😛